requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
celery-types==0.23.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
djangorestframework==3.16.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
